Move roll functions to a seperate file
